Updated extract_domain function
The previous version of this function just extracted the host (i.e., everything between the scheme e.g. https:// and the port or path e.g. /section/page123)
For example, from an URL like https://www.maps.google.com/something the previous version would extract www.maps.google.com, but most people think of google.com when they speak of the domain in that case. Of course for all the URLs in the test_data, this won't make a difference, but in the real world it will.

Feel free to adapt the code if not data.tabley enough